UART messages
Adds proper functionality to the UART from the ADC as well as buttons.

+++ App2.X/UART2.h
@@ -6,7 +6,7 @@ void XmitUART2(char, unsigned int);
 
 void __attribute__ ((interrupt, no_auto_psv)) _U2TXInterrupt(void); 
 
-void Disp2Digit(uint8_t num);
+void Disp2Digit(uint16_t num);
 void Disp2String(char*);
 
 #endif
